requirements.txt
requests==2.31.0
botbuilder-integration-aiohttp>=4.14.5
datarobotx
datarobot
textstat
tiktoken
nltk
py-readability-metrics
datarobot-mlops
datarobot-mlops-connected-client